Disentangle `tedge config` commands from CLI parsing
- Do all the parsing in `config/cli.rs`.

- Commands are defined in `config/commands/*.rs`. One (sub-)command per file.

- The implementation of the commands become easier to read
  and understand, and you can test each command independently.
