+++ tedge/src/cli/config/commands/mod.rs
@@ -1,0 +1,6 @@
+mod get;
+mod list;
+mod set;
+mod unset;
+
+pub use self::{get::*, list::*, set::*, unset::*};
